Add with default settings
Ugly but working

+++ app/Controller/RaspberriesController.php
@@ -8,6 +8,8 @@ App::uses('AppController', 'Controller');
  */
 class RaspberriesController extends AppController {
 
+public $uses = array('Raspberry','Setting');
+
 /**
  * Components
  *
@@ -49,51 +51,40 @@ class RaspberriesController extends AppController {
 		if ($this->request->is('post')) {
 			$this->Raspberry->create();
 			if ($this->Raspberry->save($this->request->data)) {
-                $filepath = '\\\\'.$this->request->data['Raspberry']['address'].'\Userdata\addon_data\service.openelec.settings\oe_settings.xml';
-
-                $oesettings = '<openelec>
-									<addon_config/>
-									<settings>
-										<system>
-											<wizard_completed>True</wizard_completed>
-										<hostname>'.$this->request->data['Raspberry']['name'].'</hostname></system>
-										<connman>
-											<wizard_completed>True</wizard_completed>
-										</connman>
-										<services>
-											<wizard_completed>True</wizard_completed>
-										</services>
-										<about>
-											<wizard_completed>True</wizard_completed>
-										</about>
-										<openelec>
-											<wizard_completed>True</wizard_completed>
-										</openelec>
-									</settings>
-								</openelec>';
-                //Génération du futur XML en DOMDocument
-				$dom = new XmlDOM();
-            	$dom->preserveWhiteSpace = false;
-				$dom->loadXML($oesettings);
-
-				$connection = $this->connexionSSH($this->request->data['Raspberry']['address'],'root','openelec');
+				$address = $this->request->data['Raspberry']['address'];
+				$files = array(
+					'guisettings' => array('Interface', $address.'\guisettings.xml'),
+					'advancedsettings' => array('Parametres Avancés', $address.'\advancedsettings.xml'),
+					'oe_settings' => array('Raspberry', $address.'\addon_data\service.openelec.settings\oe_settings.xml')
+					);
+				$connection = $this->connexionSSH($address,'root','openelec');
 				$this->execSSH($connection,'systemctl stop kodi');
-				sleep(2);
-				if ($dom->save($filepath))
-				{
-					$this->execSSH($connection,'systemctl start kodi');
-					sleep(1);
-					$this->execSSH($connection,'systemctl restart kodi');
-					sleep(1);
+				foreach ($files as $filename=>$file) {
+					$dom = new DOMDocument();
+		            $dom->preserveWhiteSpace = false;
+		            if(file_exists('./files/'.$filename.'.xml')) 
+		                {
+							$dom->load('./files/'.$filename.'.xml');
+							if ($filename == 'oe_settings')
+							{
+								$dom->getElementsByTagName("hostname")->item(0)->nodeValue = $this->request->data['Raspberry']['name'];
+							}
+							$settings = array('Setting' => array('name' => $filename, 'description' => $file[0], 'path' => $file[1], 'extension' => 'xml', 'raspberries_id' => $this->Raspberry->id));
+							$this->Setting->create();
 
-					$this->Session->setFlash(__('The raspberry has been saved'), 'flash/success');
-					$this->redirect(array('action' => 'index'));
+							if ((!$dom->save('\\\\'.$file[1].$filename.'.xml')) || (!$this->Setting->save($settings))) $this->Session->setFlash(__('Error when saving '.$filename.'. Please, try again.'), 'flash/error');
+						}
+					}
+				sleep(2);
+				$this->execSSH($connection,'systemctl start kodi');
+				sleep(1);
+				$this->execSSH($connection,'systemctl restart kodi');
+				sleep(1);
+				$this->Session->setFlash(__('The raspberry has been saved'), 'flash/success');
+				$this->redirect(array('action' => 'index'));
 				} else {
-					$this->Session->setFlash(__('The raspberry could not be saved. Please, try again.'), 'flash/error');
-				}
-			} else {
-				$this->Session->setFlash(__('The raspberry could not be saved. Please, try again.'), 'flash/error');
-			}
+							$this->Session->setFlash(__('The raspberry could not be saved. Please, try again.'), 'flash/error');
+						}
 		}
 	}
 
@@ -138,9 +129,13 @@ class RaspberriesController extends AppController {
 		if (!$this->Raspberry->exists()) {
 			throw new NotFoundException(__('Invalid raspberry'));
 		}
-		if ($this->Raspberry->delete()) {
-			$this->Session->setFlash(__('Raspberry deleted'), 'flash/success');
-			$this->redirect(array('action' => 'index'));
+		$options = array('conditions' => array('Setting.raspberries_id' => $id));
+		$this->Setting->id = $this->Raspberry->find('first', $options);
+		if ($this->Setting->delete()) {
+			if ($this->Raspberry->delete()) {
+				$this->Session->setFlash(__('Raspberry deleted'), 'flash/success');
+				$this->redirect(array('action' => 'index'));
+			}
 		}
 		$this->Session->setFlash(__('Raspberry was not deleted'), 'flash/error');
 		$this->redirect(array('action' => 'index'));
@@ -247,7 +242,49 @@ class RaspberriesController extends AppController {
 		if ($this->Raspberry->exists($id)) {
 			$options = array('conditions' => array('Raspberry.' . $this->Raspberry->primaryKey => $id));
 			$this->set('raspberry', $this->Raspberry->find('first', $options));
+			$raspberry = $this->Raspberry->find('first', $options);
+		} else {
+			// $options = array('conditions' => array('Raspberry.role' => 'master'));
+			// $this->set('raspberry', $this->Raspberry->find('first', $options));
+			$raspberry = $this->Raspberry->find('all');
 		}
+
+		if ($this->request->is('post') || $this->request->is('put')) {
+			foreach ($raspberry as $rasp) {
+				$address = $rasp['Raspberry']['address'];
+				$filepath = '\\\\'.$address.'\Userdata\\';
+				$files = array(
+					'guisettings' => array($address.'\guisettings.xml', 'Interface', 'Blabla1'),
+					'advancedsettings' => array($address.'\advancedsettings.xml', 'Parametres Avancés', 'Blabla2'),
+					'oe_settings' => array($address.'\addon_data\service.openelec.settings\oe_settings.xml', 'Raspberry', 'Blabla3')
+					);
+				foreach ($files as $filename=>$file) {
+					$dom = new DOMDocument();
+		            $dom->preserveWhiteSpace = false;
+		            if(file_exists('./files/'.$filename.'.xml')) 
+		                {
+							$dom->load('./files/'.$filename.'.xml');
+						}
+
+						$connection = $this->connexionSSH($address,'root','openelec');
+						$this->execSSH($connection,'systemctl stop kodi');
+						sleep(2);
+						if ($dom->save($filepath.$filename.'.xml'))
+						{
+							$this->execSSH($connection,'systemctl start kodi');
+							sleep(1);
+							$this->execSSH($connection,'systemctl restart kodi');
+							sleep(1);
+						} else {
+							$this->Session->setFlash(__('Error when saving '.$filename), 'flash/error');
+							exit;
+						}
+					}
+				$this->Session->setFlash(__('New file(s) has been saved'), 'flash/success');
+				$this->redirect(array('action' => 'settings', $id));
+			}
+		}
+	            
 	}
 
 }
